feat(shoppinglist-service): add auth context to controller (WIP)

+++ src/shoppinglist-service/userShoppingList/default_controller.go
@@ -2,6 +2,7 @@ package userShoppingList
 
 import (
 	"encoding/json"
+	"hsfl.de/group6/hsfl-master-ai-cloud-engineering/lib/router/middleware/auth"
 	"hsfl.de/group6/hsfl-master-ai-cloud-engineering/shoppinglist-service/userShoppingList/model"
 	"net/http"
 	"strconv"
@@ -15,31 +16,67 @@ func NewDefaultController(userShoppingListRepository Repository) *defaultControl
 	return &defaultController{userShoppingListRepository}
 }
 
-func (controller defaultController) GetList(writer http.ResponseWriter, request *http.Request) {
-	listId, err := strconv.ParseUint(request.Context().Value("listId").(string), 10, 64)
-	if err != nil {
-		http.Error(writer, err.Error(), http.StatusBadRequest)
-		return
-	}
+func (controller defaultController) GetLists(writer http.ResponseWriter, request *http.Request) {
 	userId, err := strconv.ParseUint(request.Context().Value("userId").(string), 10, 64)
 	if err != nil {
 		http.Error(writer, err.Error(), http.StatusBadRequest)
 		return
 	}
 
-	value, err := controller.userShoppingListRepository.FindByIds(userId, listId)
-	if err != nil {
-		if err.Error() == ErrorListNotFound {
-			http.Error(writer, err.Error(), http.StatusNotFound)
+	authUserId, _ := request.Context().Value("auth_userId").(uint64)
+	authUserRole, _ := request.Context().Value("auth_userRole").(int64)
+
+	values, err := controller.userShoppingListRepository.FindAllById(userId)
+
+	if authUserId == userId || authUserRole == auth.Administrator {
+		if err != nil {
+			http.Error(writer, err.Error(), http.StatusInternalServerError)
+		}
+
+		writer.Header().Set("Content-Type", "application/json")
+		err = json.NewEncoder(writer).Encode(values)
+		if err != nil {
+			http.Error(writer, err.Error(), http.StatusInternalServerError)
 			return
 		}
-		http.Error(writer, err.Error(), http.StatusInternalServerError)
+	} else {
+		writer.WriteHeader(http.StatusUnauthorized)
+	}
+}
+
+func (controller defaultController) GetList(writer http.ResponseWriter, request *http.Request) {
+	listId, err := strconv.ParseUint(request.Context().Value("listId").(string), 10, 64)
+	if err != nil {
+		http.Error(writer, err.Error(), http.StatusBadRequest)
+		return
 	}
 
-	writer.Header().Set("Content-Type", "application/json")
-	err = json.NewEncoder(writer).Encode(value)
+	userId, err := strconv.ParseUint(request.Context().Value("userId").(string), 10, 64)
 	if err != nil {
-		http.Error(writer, err.Error(), http.StatusInternalServerError)
+		http.Error(writer, err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	authUserId, _ := request.Context().Value("auth_userId").(uint64)
+	authUserRole, _ := request.Context().Value("auth_userRole").(int64)
+
+	if authUserId == userId || authUserRole == auth.Administrator {
+		value, err := controller.userShoppingListRepository.FindByIds(userId, listId)
+		if err != nil {
+			if err.Error() == ErrorListNotFound {
+				http.Error(writer, err.Error(), http.StatusNotFound)
+				return
+			}
+			http.Error(writer, err.Error(), http.StatusInternalServerError)
+		}
+
+		writer.Header().Set("Content-Type", "application/json")
+		err = json.NewEncoder(writer).Encode(value)
+		if err != nil {
+			http.Error(writer, err.Error(), http.StatusInternalServerError)
+		}
+	} else {
+		writer.WriteHeader(http.StatusUnauthorized)
 	}
 }
 
@@ -62,14 +99,21 @@ func (controller defaultController) PutList(writer http.ResponseWriter, request 
 		return
 	}
 
-	if _, err := controller.userShoppingListRepository.Update(&model.UserShoppingList{
-		Id:          listId,
-		UserId:      userId,
-		Description: requestData.Description,
-		Completed:   requestData.Checked,
-	}); err != nil {
-		writer.WriteHeader(http.StatusInternalServerError)
-		return
+	authUserId, _ := request.Context().Value("auth_userId").(uint64)
+	authUserRole, _ := request.Context().Value("auth_userRole").(int64)
+
+	if authUserId == userId || authUserRole == auth.Administrator {
+		if _, err := controller.userShoppingListRepository.Update(&model.UserShoppingList{
+			Id:          listId,
+			UserId:      userId,
+			Description: requestData.Description,
+			Completed:   requestData.Checked,
+		}); err != nil {
+			writer.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+	} else {
+		writer.WriteHeader(http.StatusUnauthorized)
 	}
 }
 
@@ -86,33 +130,21 @@ func (controller defaultController) PostList(writer http.ResponseWriter, request
 		return
 	}
 
-	if _, err := controller.userShoppingListRepository.Create(&model.UserShoppingList{
-		UserId:      userId,
-		Description: requestData.Description,
-		Completed:   false,
-	}); err != nil {
-		writer.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	writer.WriteHeader(http.StatusCreated)
-}
+	authUserId, _ := request.Context().Value("auth_userId").(uint64)
+	authUserRole, _ := request.Context().Value("auth_userRole").(int64)
 
-func (controller defaultController) GetLists(writer http.ResponseWriter, request *http.Request) {
-	userId, err := strconv.ParseUint(request.Context().Value("userId").(string), 10, 64)
-	if err != nil {
-		http.Error(writer, err.Error(), http.StatusBadRequest)
-		return
-	}
-	values, err := controller.userShoppingListRepository.FindAllById(userId)
-	if err != nil {
-		http.Error(writer, err.Error(), http.StatusInternalServerError)
-	}
-
-	writer.Header().Set("Content-Type", "application/json")
-	err = json.NewEncoder(writer).Encode(values)
-	if err != nil {
-		http.Error(writer, err.Error(), http.StatusInternalServerError)
-		return
+	if authUserId == userId || authUserRole == auth.Administrator {
+		if _, err := controller.userShoppingListRepository.Create(&model.UserShoppingList{
+			UserId:      userId,
+			Description: requestData.Description,
+			Completed:   false,
+		}); err != nil {
+			writer.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+		writer.WriteHeader(http.StatusCreated)
+	} else {
+		writer.WriteHeader(http.StatusUnauthorized)
 	}
 }
 
@@ -123,8 +155,20 @@ func (controller defaultController) DeleteList(writer http.ResponseWriter, reque
 		return
 	}
 
-	if err := controller.userShoppingListRepository.Delete(&model.UserShoppingList{Id: listId}); err != nil {
-		writer.WriteHeader(http.StatusInternalServerError)
-		return
+	list, err := controller.userShoppingListRepository.FindById(listId)
+	if err != nil {
+		http.Error(writer, err.Error(), http.StatusNotFound)
+	}
+
+	authUserId, _ := request.Context().Value("auth_userId").(uint64)
+	authUserRole, _ := request.Context().Value("auth_userRole").(int64)
+
+	if authUserId == list.UserId || authUserRole == auth.Administrator {
+		if err := controller.userShoppingListRepository.Delete(&model.UserShoppingList{Id: listId}); err != nil {
+			writer.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+	} else {
+		writer.WriteHeader(http.StatusUnauthorized)
 	}
 }
